+++ CHANGES.rst
@@ -13,6 +13,8 @@ Unreleased
 
 * Update entries whenever their :attr:`~Entry.updated` changes,
   don't compare the values. (:issue:`332`)
+* Group mutually-exclusive attributes of :class:`~.FeedUpdateIntent`
+  into its :attr:`~.FeedUpdateIntent.value` union attribute. (:issue:`332`)
   
 * Fix bug introduced in `version 3.12 <Version 3.12_>`_ causing an assertion error
   when there are multiple entries with the same id in the same feed,